Android App able to be compiled
- updated package from com.example.appcenter.sampleapp_android -> ms.appcenter.sampleapp.android
- updated dependencies to latest versions
- updated dependencies to androidx
- explicitly declare java version compatibility (better interop with Intellij IDE)
