+++ app/src/main/java/ms/appcenter/sampleapp/android/DistributeActivity.java
@@ -1,25 +1,26 @@
-package com.example.appcenter.sampleapp_android;
+package ms.appcenter.sampleapp.android;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class WelcomeActivity extends Fragment {
-    private static final String pageName = "Welcome";
+import androidx.fragment.app.Fragment;
+
+public class DistributeActivity extends Fragment {
+    private static final String pageName = "Distribute";
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         ViewGroup rootView = (ViewGroup) inflater.inflate(
-                R.layout.welcome_root, container, false);
+                R.layout.distribute_root, container, false);
         return rootView;
     }
 
-    public static WelcomeActivity newInstance() {
+    public static DistributeActivity newInstance() {
         Bundle args = new Bundle();
-        WelcomeActivity fragment = new WelcomeActivity();
+        DistributeActivity fragment = new DistributeActivity();
         fragment.setArguments(args);
         return fragment;
     }
@@ -28,3 +29,4 @@ public class WelcomeActivity extends Fragment {
         return pageName;
     }
 }
+
