+++ app/src/main/java/ms/appcenter/sampleapp/android/PushActivity.java
@@ -1,25 +1,26 @@
-package com.example.appcenter.sampleapp_android;
+package ms.appcenter.sampleapp.android;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class BuildActivity extends Fragment {
-    private static final String pageName = "Build";
+import androidx.fragment.app.Fragment;
+
+public class PushActivity extends Fragment {
+    private static final String pageName = "Push";
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         ViewGroup rootView = (ViewGroup) inflater.inflate(
-                R.layout.build_root, container, false);
+                R.layout.push_root, container, false);
         return rootView;
     }
 
-    public static BuildActivity newInstance() {
+    public static PushActivity newInstance() {
         Bundle args = new Bundle();
-        BuildActivity fragment = new BuildActivity();
+        PushActivity fragment = new PushActivity();
         fragment.setArguments(args);
         return fragment;
     }
