+++ app/src/main/java/ms/appcenter/sampleapp/android/TestActivity.java
@@ -1,10 +1,11 @@
-package com.example.appcenter.sampleapp_android;
+package ms.appcenter.sampleapp.android;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
 
 public class TestActivity extends Fragment {
     private static final String pageName = "Test";
