+++ app/src/main/java/ms/appcenter/sampleapp/android/WelcomeActivity.java
@@ -1,25 +1,26 @@
-package com.example.appcenter.sampleapp_android;
+package ms.appcenter.sampleapp.android;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class DistributeActivity extends Fragment {
-    private static final String pageName = "Distribute";
+import androidx.fragment.app.Fragment;
+
+public class WelcomeActivity extends Fragment {
+    private static final String pageName = "Welcome";
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         ViewGroup rootView = (ViewGroup) inflater.inflate(
-                R.layout.distribute_root, container, false);
+                R.layout.welcome_root, container, false);
         return rootView;
     }
 
-    public static DistributeActivity newInstance() {
+    public static WelcomeActivity newInstance() {
         Bundle args = new Bundle();
-        DistributeActivity fragment = new DistributeActivity();
+        WelcomeActivity fragment = new WelcomeActivity();
         fragment.setArguments(args);
         return fragment;
     }
@@ -28,4 +29,3 @@ public class DistributeActivity extends Fragment {
         return pageName;
     }
 }
-
