+++ app/src/test/java/ms/appcenter/sampleapp/android/ExampleUnitTest.java
@@ -1,8 +1,8 @@
-package com.example.appcenter.sampleapp_android;
+package ms.appcenter.sampleapp.android;
 
 import org.junit.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 /**
  * Example local unit test, which will execute on the development machine (host).
@@ -11,7 +11,7 @@ import static org.junit.Assert.*;
  */
 public class ExampleUnitTest {
     @Test
-    public void addition_isCorrect() throws Exception {
+    public void addition_isCorrect() {
         assertEquals(4, 2 + 2);
     }
 }
